pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "receipt-ocr"
version = "0.1.0"
description = "Extract structured data from receipt images using Gemini AI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "IndieKit", email = "hi@indiekit.ai"}
]
keywords = ["receipt", "ocr", "gemini", "ai", "invoice", "extraction"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.0.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
receipt-ocr = "receipt_ocr.cli:main"

[project.urls]
Homepage = "https://github.com/indiekitai/receipt-ocr"
Repository = "https://github.com/indiekitai/receipt-ocr"
Issues = "https://github.com/indiekitai/receipt-ocr/issues"

[tool.setuptools.packages.find]
include = ["receipt_ocr*"]
